+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * **[CHANGE]** Filters and sorts will be now reset after each page visit (requested by [yeah idk](https://app.feedbacky.net/i/reseting-the-filtering-and-sorting-types.2006))
 * **[CHANGE]** Default sort for comments is now Newest
 * **[FIX]** Google users couldn't log-in for the first time if their avatar URL length exceeded 255 characters
+* **[FIX]** Moderators couldn't be removed
 
 ### 1.0.0.alpha.3
 * **[FEATURE]** Added anonymous voting feature, toggleable in board admin panel (requested at [Feedbacky](https://app.feedbacky.net/i/anonymous-voting.730))